R version 4.3.0 had a breaking change.
It included a 'fix' for ns and bs that affected the results to become far worse. The code in suggest_knotcount has now been rewritten, and should work correctly. However, the performance may or may not be a bit affected.

See https://stat.ethz.ch/pipermail/r-announce/2023/000691.html about the change in R 4.3.0 that was based on PR#18442 by Ben Bolker